Feature: Added automatic mapping output
This code automatically dumps the list of keyboard mappings to
`~/.config/nvim/mappings` to make it easier for users to see which keys
are mapped.

This is better than just using `:map` or `:nmap` because the output can
be filtered, edited, and commented. I use this all the time.

Also added `.gitkeep` to make sure the directory is present.