Bug fix: Added missing config for nvim-notify
I was getting the following message when using nvim-notify:

Highlight group 'NotifyBackground' has no background highlight
Please provide an RGB hex value or highlight group with a background
value for 'background_colour' option.
This is the colour that will be used for 100% transparency.
```lua
require("notify").setup({
  background_colour = "#000000",
})
```
Defaulting to #000000
Press ENTER or type command to continue

Adding the configuration to the plugin_setup module fixes this.