refactor: Move nvim-cmp section up and remove org-mode
The nvim-cmp section had to be moved higher in the file. Not entirely
sure why but I got in a state where completion wouldn't work anymore and
this fixed it.

Also, get rid of org-mode because it is obsoleted by other tools like
taskwarrior, taskell, and time tracking software that I have built
myself.